requirements.txt
APScheduler==3.11.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
Flask==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
ping3==5.1.3
ply==3.11
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodomex==3.23.0
pysmi==1.6.2
pysnmp==4.4.12
PyYAML==6.0.2
requests==2.32.4
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
Werkzeug==3.1.3
